fix build

+++ src/util/log/Logger.cpp
@@ -138,8 +138,8 @@ LogService::init(config::ClioConfigDefinition const& config)
         auto const rotationPeriod = config.get<uint32_t>("log_rotation_hour_interval");
 
         // the below are taken from user in MB, but boost::log::add_file_log needs it to be in bytes
-        auto const rotationSize = mbToBytes(config.get<uint64_t>("log_rotation_size"));
-        auto const dirSize = mbToBytes(config.get<uint64_t>("log_directory_max_size"));
+        auto const rotationSize = mbToBytes(config.get<uint32_t>("log_rotation_size"));
+        auto const dirSize = mbToBytes(config.get<uint32_t>("log_directory_max_size"));
         auto fileSink = boost::log::add_file_log(
             keywords::file_name = dirPath / "clio.log",
             keywords::target_file_name = dirPath / "clio_%Y-%m-%d_%H-%M-%S.log",
